Updated examples and README
