+++ examples/1-animating-lines.md
@@ -1,7 +1,7 @@
 # Final Code
 
 ```
-var line = Line({ form: w.Lines.SlopePoint, slope: 1, point: {x: 0, y: 0}, color: w.Colors.Green(), thickness: 5 })
+var line = Line({ form: w.Lines.SlopePoint, slope: 1, point: {x: 0, y: 0}, color: Colors.green, thickness: 5 })
 
 Create(line, { duration: 10 })
 ```
