+++ examples/2-animating-points.md
@@ -1,7 +1,7 @@
 # Final Code
 
 ```
-var point = Point({ x: 500, y: 500, color: Colors.Red(), size: 100 })
+var point = Point({ x: 500, y: 500, color: Colors.red, size: 100 })
 
 Create(point)
 ```
