+++ examples/4-plotting-points.md
@@ -6,5 +6,5 @@ var plane = NumberPlane({ grid: true })
 FadeIn(plane)
 await wait(1000)
 
-plane.point({ x: -1, y: 1, transition: Transitions.Create, color: Colors.Orange(), size: 10 })
+Create(plane.point({ x: -1, y: 1 color: Colors.orange, size: 10 }))
 ```
