+++ examples/5-plotting-functions.md
@@ -6,7 +6,7 @@ var plane = NumberPlane({ grid: true })
 FadeIn(plane)
 await wait(1000)
 
-plane.plot({ definition: 'y = sin(x)', color: Colors.Green(), thickness: 5, transition: Transitions.Create })
+Create(plane.plot({ definition: 'y = sin(x)', color: Colors.green, thickness: 5 }))
 ```
 
 ### Currently, only explicity functions are supported by AnimWeb. Support for Implicit and Parametric functions will come soon.
