+++ examples/6-tangent-to-function.md
@@ -6,10 +6,10 @@ var plane = NumberPlane({ grid: true })
 FadeIn(plane)
 await wait(1000)
 
-var cosx = await plane.plot({ definition: 'y = cos(x)', color: Colors.Green(), thickness: 5, transition: Transitions.FadeIn })
+var cosx = FadeIn(plane.plot({ definition: 'y = cos(x)', color: Colors.green, thickness: 5 }))
 await wait()
 
-var tangent = await cosx.addAnchorLine({ x: 0, color: Colors.Orange(), thickness: 3, transition: Transitions.Create })
+var tangent = Create(cosx.addTangent({ x: 0, color: Colors.orange, thickness: 3 }))
 await wait(1000)
 
 tangent.moveTo({ x: Math.PI })
