+++ examples/8-anchor-points.md
@@ -6,11 +6,11 @@ Create(plane)
 
 await wait(3000)
 
-var sinx = await plane.plot({ definition: 'y = sin(x)', transition: Transitions.Create, color: Colors.Green(), thickness: 5 })
+var sinx = Create(plane.plot({ definition: 'y = sin(x)', color: Colors.green, thickness: 5 }))
 
 await wait()
 
-var point = await sinx.addAnchorPoint({ x: 0, color: Colors.Orange(), size: 10, transition: Transitions.Create })
+var point = Create(sinx.addAnchorPoint({ x: 0, color: Colors.orange, size: 10 }))
 
 await wait(1000)
