+++ pages/login.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+<span>login page - work in progress</span>
+</template>
+
+<style>
+
+</style>
